tests/wpt/tests/html/canvas/element/transformations/2d.transformation.scale.multiple.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.transformation.scale.multiple</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.transformation.scale.multiple</h1>
<p class="desc">Multiple scale()s combine</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Multiple scale()s combine");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  ctx.scale(Math.sqrt(2), Math.sqrt(2));
  ctx.scale(Math.sqrt(2), Math.sqrt(2));
  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 50, 25);
  _assertPixel(canvas, 90,40, 0,255,0,255);

});
</script>
